move utils used in EV from DVRP to common
