+++ contribs/common/src/main/java/org/matsim/contrib/common/collections/PartialSort.java
@@ -1,9 +1,9 @@
-/* *********************************************************************** *
+/*
+ * *********************************************************************** *
  * project: org.matsim.*
- *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2014 by the members listed in the COPYING,        *
+ * copyright       : (C) 2021 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -15,9 +15,10 @@
  *   (at your option) any later version.                                   *
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
- * *********************************************************************** */
+ * *********************************************************************** *
+ */
 
-package org.matsim.contrib.util;
+package org.matsim.contrib.common.collections;
 
 import java.util.Comparator;
 import java.util.List;
@@ -36,6 +37,8 @@ import com.google.common.base.Preconditions;
  * More info: <a href="http://en.wikipedia.org/wiki/Partial_sorting">Partial sorting</a>
  *
  * @param <T>
+ *
+ * @author michal (Michal Maciejewski)
  */
 public class PartialSort<T> {
 	public static <T> List<T> kSmallestElements(int k, Stream<T> elements, Comparator<T> comparator) {
