+++ contribs/common/src/main/java/org/matsim/contrib/common/collections/SpecificationContainer.java
@@ -3,7 +3,7 @@
  * project: org.matsim.*
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2019 by the members listed in the COPYING,        *
+ * copyright       : (C) 2021 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -18,7 +18,7 @@
  * *********************************************************************** *
  */
 
-package org.matsim.contrib.util;
+package org.matsim.contrib.common.collections;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
