+++ contribs/common/src/main/java/org/matsim/contrib/common/util/DistanceUtils.java
@@ -3,7 +3,7 @@
  * project: org.matsim.*
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2019 by the members listed in the COPYING,        *
+ * copyright       : (C) 2021 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -18,37 +18,36 @@
  * *********************************************************************** *
  */
 
-package org.matsim.contrib.dvrp.fleet;
+package org.matsim.contrib.common.util;
 
-import java.util.Map;
-
-import org.matsim.api.core.v01.Id;
-import org.matsim.contrib.common.collections.SpecificationContainer;
+import org.matsim.api.core.v01.BasicLocation;
+import org.matsim.api.core.v01.Coord;
 
 /**
  * @author Michal Maciejewski (michalm)
  */
-public final class FleetSpecificationImpl implements FleetSpecification {
-	private final SpecificationContainer<DvrpVehicle, DvrpVehicleSpecification> container = new SpecificationContainer<>();
-
-	@Override
-	public Map<Id<DvrpVehicle>, DvrpVehicleSpecification> getVehicleSpecifications() {
-		return container.getSpecifications();
+public class DistanceUtils {
+	public static double calculateDistance(BasicLocation fromLocation, BasicLocation toLocation) {
+		return calculateDistance(fromLocation.getCoord(), toLocation.getCoord());
 	}
 
-	@Override
-	public void addVehicleSpecification(DvrpVehicleSpecification specification) {
-		container.addSpecification(specification);
+	public static double calculateSquaredDistance(BasicLocation fromLocation, BasicLocation toLocation) {
+		return calculateSquaredDistance(fromLocation.getCoord(), toLocation.getCoord());
 	}
 
-	@Override
-	public void replaceVehicleSpecification(DvrpVehicleSpecification specification) {
-		container.replaceSpecification(specification);
+	/**
+	 * @return distance (for distance-based comparison/sorting, consider using the squared distance)
+	 */
+	public static double calculateDistance(Coord fromCoord, Coord toCoord) {
+		return Math.sqrt(calculateSquaredDistance(fromCoord, toCoord));
 	}
 
-	@Override
-	public void removeVehicleSpecification(Id<DvrpVehicle> vehicleId) {
-		container.removeSpecification(vehicleId);
+	/**
+	 * @return SQUARED distance (to avoid unnecessary Math.sqrt() calls when comparing distances)
+	 */
+	public static double calculateSquaredDistance(Coord fromCoord, Coord toCoord) {
+		double deltaX = toCoord.getX() - fromCoord.getX();
+		double deltaY = toCoord.getY() - fromCoord.getY();
+		return deltaX * deltaX + deltaY * deltaY;
 	}
 }
-
