+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/rebalancing/plusOne/FastHeuristicLinkBasedRelocationCalculator.java
@@ -9,7 +9,7 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.contrib.drt.optimizer.rebalancing.RebalancingStrategy.Relocation;
 import org.matsim.contrib.dvrp.fleet.DvrpVehicle;
 import org.matsim.contrib.dvrp.schedule.Schedules;
-import org.matsim.contrib.util.distance.DistanceUtils;
+import org.matsim.contrib.common.util.DistanceUtils;
 
 public class FastHeuristicLinkBasedRelocationCalculator implements LinkBasedRelocationCalculator {
