+++ contribs/dvrp/pom.xml
@@ -28,6 +28,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.matsim.contrib</groupId>
+			<artifactId>common</artifactId>
+			<version>14.0-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>org.matsim.contrib</groupId>
 			<artifactId>locationchoice</artifactId>
 			<version>14.0-SNAPSHOT</version>
 		</dependency>
@@ -39,10 +44,6 @@
 		<dependency>
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-math3</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>com.opencsv</groupId>
-			<artifactId>opencsv</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>one.util</groupId>
