+++ contribs/ev/src/main/java/org/matsim/contrib/ev/temperature/TemperatureChangeModule.java
@@ -21,13 +21,11 @@ package org.matsim.contrib.ev.temperature;/*
  * created by jbischoff, 15.08.2018
  */
 
-import org.matsim.contrib.dynagent.run.DynQSimConfigConsistencyChecker;
 import org.matsim.core.controler.AbstractModule;
 
 public class TemperatureChangeModule extends AbstractModule {
     @Override
     public void install() {
-		new DynQSimConfigConsistencyChecker().checkConsistency(getConfig());
 		addMobsimListenerBinding().to(TemperatureManager.class).asEagerSingleton();
 		bind(TemperatureServiceImpl.class).asEagerSingleton();
 		bind(TemperatureService.class).to(TemperatureServiceImpl.class);
