+++ contribs/ev/src/main/java/org/matsim/contrib/ev/temperature/TemperatureServiceImpl.java
@@ -29,7 +29,7 @@ import javax.inject.Inject;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
-import org.matsim.contrib.util.distance.DistanceUtils;
+import org.matsim.contrib.common.util.DistanceUtils;
 
 public class TemperatureServiceImpl implements TemperatureService, TemperatureChangeEventHandler {
 	private final Network network;
